Move storage drivers into their own packages
